requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
python-dotenv==1.0.1
astronomer-cosmos==1.8.0
apache-airflow-providers-snowflake==6.0.0
dbt-snowflake==1.8.4